pyproject.toml
[project]
name = "firstcommand-prototype"
version = "0.1.0"
description = "Prototype for the FirstCommand financial onboarding use case."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "openai-agents>=0.12.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
